Added logic to write variable out to script scope if part of the executed line
If we were re-executing a line that writes something to a variable, and the remainder of the script being ran depends on that value, it would not be populated as this is re-executed in the function scope only. This change uses regex to identify if the executing line starts with '$<any letters/numbers> =' and if so, changes $ to $script: so that the value is written to the script scope and accessible to the rest of the script. Also changed the & shorthand to Invoke-Command for readability